// src/constants/chain.enum.ts
/**
 * Vendored from `web3compassapi/src/constants/chain.enum.ts`. Keep in sync.
 */
export enum ChainEnum {
  ETH = 'ethereum',
  POLYGON = 'polygon',
  BSC = 'bsc',
  SOLANA = 'solana',
}

export enum DnsTypeEnum {
  ENS = 'ens',
  TOMI = 'tomi',
  UNS = 'uns',
  BNB = 'bnb',
  SNS = 'sns',
}

export const ENS_NAME_WRAPPER_ADDRESS =
  '0xd4416b13d2b3a9abae7acd5d6c2bbdbe25686401';