// src/constants/migration-type.enum.ts
/**
 * Vendored from `web3compassapi/src/constants/migration-type.enum.ts`. Keep in sync.
 */
export enum MigrationTypeEnum {
  DEFAULT = 'DEFAULT',
  ENS_REGISTER = 'ENS_REGISTER',
  ENS_UPDATE = 'ENS_UPDATE',
  ENS_RESOLVER_UPDATE = 'ENS_RESOLVER_UPDATE',

  TOMI_REGISTER = 'TOMI_REGISTER',
  TOMI_UPDATE = 'TOMI_UPDATE',
  TOMI_RESOLVER_UPDATE = 'TOMI_RESOLVER_UPDATE',

  BNB_REGISTER = 'BNB_REGISTER',
  BNB_UPDATE = 'BNB_UPDATE',
  BNB_RESOLVER_UPDATE = 'BNB_RESOLVER_UPDATE',

  SNS_REGISTER = 'SNS_REGISTER',
  SNS_RECORDS_V2_UPDATE = 'SNS_RECORDS_V2_UPDATE',
  SNS_BACKFILL = 'SNS_BACKFILL',
}